Warn player when they've messed up
Many times I've have made an early mistake that makes
the game unfinishable, this is quite annoying.

With this change set, if the player places a wrong number
and then continues the game (by playing another number),
we simply alert that this has happened, and that the game
is pointless to be carried on.